Update README.md
the banner is making the page looking bad because of Broken Relative Resource Reference.. and it is taking me a loong time to solve it. Silly me!!! hope this time it is done for. Wasting time on one thing day after another and couldn't able to fix it.. It is not me awhhhh!! hopefully this time 🤞.